Removed registration logic from DefaultServiceProvider and renamed it to ActivatorServiceProvider
